Remove gulp because of security vulnerabilities
