+++ .gitignore
@@ -17,9 +17,6 @@ coverage
 # nyc test coverage
 .nyc_output
 
-# Grunt intermediate storage (http://gruntjs.com/creating-plugins#storing-task-files)
-.grunt
-
 # node-waf configuration
 .lock-wscript
 
@@ -47,3 +44,4 @@ dist
 
 # The only allowed .js file
 !copyTestAssets.js
+!Gruntfile.js
