--- .travis.yml
@@ -1,26 +1,0 @@
-sudo: required
-language: node_js
-services:
-  - docker
-  - mongodb
-  - postgresql
-node_js:
-  - 10
-  - 12
-  - 14
-install:
-  - npm install
-before_script:
-  - sudo service mysql stop
-  - docker run -e 'ACCEPT_EULA=Y' -e 'SA_PASSWORD=yourStrong(!)Password' -p 14330:1433 -d mcr.microsoft.com/mssql/server:2017-CU8-ubuntu
-script:
-  - gulp init
-  - gulp test
-env:
-  global:
-  - POSTGRES_USER=postgres
-  - POSTGRES_PASSWORD=test
-  - POSTGRES_DB=postgres
-  - POSTGRES_HOST=127.0.0.1
-
-
