+++ package.json
@@ -4,16 +4,15 @@
   "description": "**diagnostic-channel** provides a shared channel for handling instrumentation\r and diagnostic messages in Node.js apps. Instrumentation patchers and module\r authors can publish messages to this channel by calling `channel.publish(...)`,\r and APM providers and other tools can subscribe to those messages by calling\r `channel.subscribe(...)`. Subscribers can transform the original generic message\r as needed in their handlers and relay the message on to storage and monitoring\r systems.",
   "main": "gulpfile.js",
   "scripts": {
-    "postinstall": "gulp init",
-    "build": "gulp build-main",
+    "postinstall": "grunt init",
+    "build": "grunt build",
     "lint": "eslint ./ --fix",
-    "clean": "gulp clean",
-    "test": "gulp test"
+    "clean": "grunt clean",
+    "test": "grunt test"
   },
   "dependencies": {
     "@azure/opentelemetry-instrumentation-azure-sdk": "^1.0.0-beta.1",
-    "@types/node": "^8.0.0",
-    "gulp": "^4.0.2"
+    "@types/node": "^8.0.0"
   },
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "^5.11.0",
@@ -23,7 +22,8 @@
     "eslint-plugin-import": "^2.23.4",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^5.1.0",
-    "eslint-plugin-security": "^1.4.0"
+    "eslint-plugin-security": "^1.4.0",
+    "grunt": "^1.6.1"
   },
   "repository": {
     "type": "git",
