+++ src/diagnostic-channel/package.json
