+++ src/publisher-legacy-tests/mongo3.2.7/package.json
