+++ src/publisher-legacy-tests/winston2/package.json
